@@ -4,11 +4,11 @@
     "build": "ng build"
   },
   "peerDependencies": {
-    "@angular/cdk": "^18.0.0",
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@angular/material": "^18.0.0",
-    "@hsi/app-dev-kit": "^1.1.0",
+    "@angular/cdk": "^19.2.6",
+    "@angular/common": "^19.2.2",
+    "@angular/core": "^19.2.2",
+    "@angular/material": "^19.2.6",
+    "@hsi/app-dev-kit": "^1.2.0",
     "d3": "^7.8.5",
     "lodash-es": "^4.17.21",
     "rxjs": "^7.5.0"
