@@ -4,10 +4,10 @@
     "build": "ng build"
   },
   "peerDependencies": {
-    "@angular/cdk": "^18.0.0",
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@hsi/app-dev-kit": "^1.1.0",
+    "@angular/cdk": "^19.2.6",
+    "@angular/common": "^19.2.2",
+    "@angular/core": "^19.2.2",
+    "@hsi/app-dev-kit": "^1.2.0",
     "csstype": "^3.1.3",
     "d3": "^7.6.1",
     "file-saver": "^2.0.5",
