@@ -48,16 +46,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.2.1",
-    "@angular/cdk": "^18.2.1",
-    "@angular/common": "^18.2.1",
-    "@angular/compiler": "^18.2.1",
-    "@angular/core": "^18.2.1",
-    "@angular/forms": "^18.2.1",
-    "@angular/material": "^18.2.1",
-    "@angular/platform-browser": "^18.2.1",
-    "@angular/platform-browser-dynamic": "^18.2.1",
-    "@angular/router": "^18.2.1",
+    "@angular/animations": "19.2.2",
+    "@angular/cdk": "^19.2.6",
+    "@angular/common": "19.2.2",
+    "@angular/compiler": "19.2.2",
+    "@angular/core": "19.2.2",
+    "@angular/forms": "19.2.2",
+    "@angular/material": "^19.2.6",
+    "@angular/platform-browser": "19.2.2",
+    "@angular/platform-browser-dynamic": "19.2.2",
+    "@angular/router": "19.2.2",
     "copyfiles": "^2.4.1",
     "csstype": "^3.1.3",
     "d3": "^7.8.5",
@@ -88,60 +86,60 @@
     "unified": "^11.0.5",
     "xmldoc": "^1.2.0",
     "yaml": "^2.1.1",
-    "zone.js": "^0.14.7"
+    "zone.js": "0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.1",
-    "@angular-devkit/core": "^18.2.1",
-    "@angular-devkit/schematics": "^18.2.1",
-    "@angular-eslint/builder": "18.3.0",
-    "@angular-eslint/eslint-plugin": "18.3.0",
-    "@angular-eslint/eslint-plugin-template": "18.3.0",
-    "@angular-eslint/schematics": "18.3.0",
-    "@angular-eslint/template-parser": "18.3.0",
-    "@angular/cli": "^18.2.1",
-    "@angular/compiler-cli": "^18.2.1",
+    "@angular-devkit/build-angular": "19.2.3",
+    "@angular-devkit/core": "19.2.3",
+    "@angular-devkit/schematics": "19.2.3",
+    "@angular-eslint/builder": "19.2.0",
+    "@angular-eslint/eslint-plugin": "19.2.0",
+    "@angular-eslint/eslint-plugin-template": "19.2.0",
+    "@angular-eslint/schematics": "19.2.0",
+    "@angular-eslint/template-parser": "19.2.0",
+    "@angular/cli": "~19.2.0",
+    "@angular/compiler-cli": "19.2.2",
     "@commitlint/cli": "^18.0.0",
     "@commitlint/config-angular": "^18.0.0",
     "@compodoc/compodoc": "^1.1.19",
     "@cypress/webpack-dev-server": "^3.8.0",
     "@jscutlery/semver": "^5.3.1",
-    "@nx/angular": "19.7.3",
-    "@nx/cypress": "^19.7.3",
-    "@nx/workspace": "19.7.3",
+    "@nx/angular": "20.6.2",
+    "@nx/cypress": "20.6.2",
+    "@nx/workspace": "20.6.2",
     "@percy/cli": "^1.8.1",
     "@percy/cypress": "^3.1.2",
-    "@schematics/angular": "^18.2.1",
+    "@schematics/angular": "19.2.3",
     "@types/d3": "^7.4.3",
     "@types/geojson": "^7946.0.10",
     "@types/hast": "^3.0.4",
     "@types/jasmine": "~4.0.0",
     "@types/topojson-client": "^3.1.1",
-    "@typescript-eslint/eslint-plugin": "7.11.0",
-    "@typescript-eslint/parser": "7.11.0",
-    "@typescript-eslint/utils": "^8.5.0",
+    "@typescript-eslint/eslint-plugin": "^8.26.1",
+    "@typescript-eslint/parser": "^8.26.1",
+    "@typescript-eslint/utils": "^8.26.1",
     "cypress": "^13.13.0",
-    "cypress-real-events": "^1.13.0",
+    "cypress-real-events": "^1.14.0",
     "eslint": "^8.57.0",
-    "eslint-config-prettier": "^8.10.0",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-prettier": "^5.2.3",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^9.0.0",
-    "jasmine-core": "~4.1.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
+    "jasmine-core": "^5.6.0",
+    "karma": "^6.4.4",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-coverage": "^2.2.1",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
     "karma-junit-reporter": "^2.0.1",
     "local-cypress": "^1.2.6",
     "minimist": "^1.2.8",
-    "ng-packagr": "^18.2.1",
+    "ng-packagr": "19.2.0",
     "npm-run-all": "^4.1.5",
-    "nx": "19.7.3",
-    "prettier": "^3.3.3",
+    "nx": "20.6.2",
+    "prettier": "^3.5.3",
     "prettier-eslint": "^16.3.0",
     "tsx": "^4.17.0",
-    "typescript": "~5.4.5"
+    "typescript": "5.7.3"
   }
 }
