@@ -15,12 +15,10 @@ import {
   VicDotsModule,
   VicXQuantitativeAxisConfig,
   VicXQuantitativeAxisConfigBuilder,
-  VicXQuantitativeAxisModule,
+  VicXyAxisModule,
   VicXyBackgroundModule,
-  VicXyChartModule,
   VicYQuantitativeAxisConfig,
   VicYQuantitativeAxisConfigBuilder,
-  VicYQuantitativeAxisModule,
 } from '@hsi/viz-components';
 
 interface Datum {
@@ -48,15 +46,12 @@ const data: Datum[] = [
 
 @Component({
   selector: 'app-dots-scatterplot-example',
-  standalone: true,
   imports: [
     CommonModule,
     VicChartModule,
     VicDotsModule,
-    VicXyChartModule,
     VicXyBackgroundModule,
-    VicXQuantitativeAxisModule,
-    VicYQuantitativeAxisModule,
+    VicXyAxisModule,
   ],
   templateUrl: './dots-scatterplot-example.component.html',
   styleUrl: './dots-scatterplot-example.component.scss',
