@@ -1,11 +1,8 @@
+/* eslint-disable @angular-eslint/prefer-standalone */
 import { Component, Input } from '@angular/core';
 import { beforeEach, cy, describe, expect, it } from 'local-cypress';
 import { BarsConfig, VicBarsConfigBuilder, VicBarsModule } from '../../bars';
-import {
-  VicChartConfigBuilder,
-  VicChartModule,
-  VicXyChartModule,
-} from '../../charts';
+import { VicChartConfigBuilder, VicChartModule } from '../../charts';
 import {
   LinesConfig,
   VicLinesConfigBuilder,
@@ -18,16 +15,13 @@ import {
 import { VicXyBackgroundModule } from '../../xy-background';
 import { VicXOrdinalAxisConfigBuilder } from '../x-ordinal/x-ordinal-axis-builder';
 import { VicXOrdinalAxisConfig } from '../x-ordinal/x-ordinal-axis-config';
-import { VicXOrdinalAxisModule } from '../x-ordinal/x-ordinal-axis.module';
 import { VicXQuantitativeAxisConfigBuilder } from '../x-quantitative/x-quantitative-axis-builder';
 import { VicXQuantitativeAxisConfig } from '../x-quantitative/x-quantitative-axis-config';
-import { VicXQuantitativeAxisModule } from '../x-quantitative/x-quantitative-axis.module';
+import { VicXyAxisModule } from '../xy-axis.module';
 import { VicYOrdinalAxisConfigBuilder } from '../y-ordinal/y-ordinal-axis-builder';
 import { VicYOrdinalAxisConfig } from '../y-ordinal/y-ordinal-axis-config';
-import { VicYOrdinalAxisModule } from '../y-ordinal/y-ordinal-axis.module';
 import { VicYQuantitativeAxisConfigBuilder } from '../y-quantitative-axis/y-quantitative-axis-builder';
 import { VicYQuantitativeAxisConfig } from '../y-quantitative-axis/y-quantitative-axis-config';
-import { VicYQuantitativeAxisModule } from '../y-quantitative-axis/y-quantitative-axis.module';
 
 const axisTickTextWaitTime = 2000;
 
@@ -55,6 +49,7 @@ const data = ContinentPopulationNumYearData;
     </vic-xy-chart>
   `,
   styles: [],
+  standalone: false,
 })
 class TestZeroAxisLinesComponent<Datum> {
   @Input() linesConfig: LinesConfig<Datum>;
@@ -70,10 +65,8 @@ class TestZeroAxisLinesComponent<Datum> {
 const linesImports = [
   VicChartModule,
   VicLinesModule,
-  VicXQuantitativeAxisModule,
-  VicYQuantitativeAxisModule,
+  VicXyAxisModule,
   VicXyBackgroundModule,
-  VicXyChartModule,
 ];
 
 function mountZeroAxisLinesComponent(
@@ -177,6 +170,7 @@ describe('Domain lines positioning, two quant dimensions', () => {
     </vic-xy-chart>
   `,
   styles: [],
+  standalone: false,
 })
 class TestZeroAxisHorizontalBarsComponent<Datum> {
   @Input() barsConfig: BarsConfig<Datum, string>;
@@ -192,10 +186,8 @@ class TestZeroAxisHorizontalBarsComponent<Datum> {
 const horizontalBarsImports = [
   VicChartModule,
   VicBarsModule,
-  VicXQuantitativeAxisModule,
-  VicYOrdinalAxisModule,
+  VicXyAxisModule,
   VicXyBackgroundModule,
-  VicXyChartModule,
 ];
 
 function mountZeroAxisHorizontalBarsComponent(
@@ -304,6 +296,7 @@ describe('Domain lines positioning, one quant, one ordinal dimension - horizonta
     </vic-xy-chart>
   `,
   styles: [],
+  standalone: false,
 })
 class TestZeroAxisVerticalBarsComponent<Datum> {
   @Input() barsConfig: BarsConfig<Datum, string>;
@@ -319,10 +312,8 @@ class TestZeroAxisVerticalBarsComponent<Datum> {
 const verticalBarsImports = [
   VicChartModule,
   VicBarsModule,
-  VicYQuantitativeAxisModule,
-  VicXOrdinalAxisModule,
+  VicXyAxisModule,
   VicXyBackgroundModule,
-  VicXyChartModule,
 ];
 
 function mountZeroAxisVerticalBarsComponent(
