@@ -1,16 +1,13 @@
 export * from './x-ordinal/x-ordinal-axis-builder';
 export * from './x-ordinal/x-ordinal-axis-config';
 export * from './x-ordinal/x-ordinal-axis.component';
-export * from './x-ordinal/x-ordinal-axis.module';
 export * from './x-quantitative/x-quantitative-axis-builder';
 export * from './x-quantitative/x-quantitative-axis-config';
 export * from './x-quantitative/x-quantitative-axis.component';
-export * from './x-quantitative/x-quantitative-axis.module';
+export * from './xy-axis.module';
 export * from './y-ordinal/y-ordinal-axis-builder';
 export * from './y-ordinal/y-ordinal-axis-config';
 export * from './y-ordinal/y-ordinal-axis.component';
-export * from './y-ordinal/y-ordinal-axis.module';
 export * from './y-quantitative-axis/y-quantitative-axis-builder';
 export * from './y-quantitative-axis/y-quantitative-axis-config';
 export * from './y-quantitative-axis/y-quantitative-axis.component';
-export * from './y-quantitative-axis/y-quantitative-axis.module';
