@@ -1,3 +1,4 @@
+/* eslint-disable @angular-eslint/prefer-standalone */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Component, Input, OnInit } from '@angular/core';
 import { extent } from 'd3';
@@ -7,12 +8,11 @@ import { VicBarsModule } from '../../bars/bars.module';
 import { VicBarsConfigBuilder } from '../../bars/config/bars-builder';
 import { BarsOptions } from '../../bars/config/bars-options';
 import { ChartConfig, VicChartConfigBuilder } from '../../charts';
-import { VicChartModule } from '../../charts/chart/chart.module';
-import { VicXyChartModule } from '../../charts/xy-chart/xy-chart.module';
+import { VicChartModule } from '../../charts/chart.module';
 import { VicXyBackgroundModule } from '../../xy-background';
 import { VicXQuantitativeAxisConfigBuilder } from '../x-quantitative/x-quantitative-axis-builder';
 import { VicXQuantitativeAxisConfig } from '../x-quantitative/x-quantitative-axis-config';
-import { VicXQuantitativeAxisModule } from '../x-quantitative/x-quantitative-axis.module';
+import { VicXyAxisModule } from '../xy-axis.module';
 
 // Cypress will get the tick elements before d3 has set the text value of the elements,
 // because d3 creates the elements and sets the text value in a transition).
@@ -35,6 +35,7 @@ const tickTextSelector = '.vic-axis-x-quantitative .tick text';
     </vic-xy-chart>
   `,
   styles: [],
+  standalone: false,
   providers: [VicChartConfigBuilder],
 })
 class TestXQuantitativeAxisComponent implements OnInit {
@@ -60,8 +61,7 @@ describe('it correctly sets ticks', () => {
   const imports = [
     VicChartModule,
     VicBarsModule,
-    VicXQuantitativeAxisModule,
-    VicXyChartModule,
+    VicXyAxisModule,
     VicXyBackgroundModule,
   ];
   beforeEach(() => {
@@ -160,8 +160,7 @@ describe('integer formatted ticks', () => {
   const imports = [
     VicChartModule,
     VicBarsModule,
-    VicXQuantitativeAxisModule,
-    VicXyChartModule,
+    VicXyAxisModule,
     VicXyBackgroundModule,
   ];
   beforeEach(() => {
@@ -328,8 +327,7 @@ describe('float formatted ticks', () => {
   const imports = [
     VicChartModule,
     VicBarsModule,
-    VicXQuantitativeAxisModule,
-    VicXyChartModule,
+    VicXyAxisModule,
     VicXyBackgroundModule,
   ];
   beforeEach(() => {
@@ -509,8 +507,7 @@ describe('percent formatted ticks', () => {
   const imports = [
     VicChartModule,
     VicBarsModule,
-    VicXQuantitativeAxisModule,
-    VicXyChartModule,
+    VicXyAxisModule,
     VicXyBackgroundModule,
   ];
   beforeEach(() => {
@@ -675,8 +672,7 @@ describe('grid lines', () => {
   const imports = [
     VicChartModule,
     VicBarsModule,
-    VicXQuantitativeAxisModule,
-    VicXyChartModule,
+    VicXyAxisModule,
     VicXyBackgroundModule,
   ];
   beforeEach(() => {
