@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { BarsComponent } from './bars.component';
 import { BarsClickDirective } from './events/bars-click.directive';
@@ -7,14 +6,13 @@ import { BarsHoverDirective } from './events/bars-hover.directive';
 import { BarsInputEventDirective } from './events/bars-input-event.directive';
 
 @NgModule({
-  declarations: [
+  imports: [
     BarsComponent,
+    BarsClickDirective,
     BarsHoverDirective,
     BarsHoverMoveDirective,
     BarsInputEventDirective,
-    BarsClickDirective,
   ],
-  imports: [CommonModule],
   exports: [
     BarsComponent,
     BarsHoverDirective,
