@@ -1,3 +1,4 @@
+/* eslint-disable @angular-eslint/prefer-standalone */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Component, Input } from '@angular/core';
 import {
@@ -7,17 +8,13 @@ import {
   VicChartModule,
   VicXOrdinalAxisConfig,
   VicXOrdinalAxisConfigBuilder,
-  VicXOrdinalAxisModule,
   VicXQuantitativeAxisConfig,
   VicXQuantitativeAxisConfigBuilder,
-  VicXQuantitativeAxisModule,
-  VicXyChartModule,
+  VicXyAxisModule,
   VicYOrdinalAxisConfig,
   VicYOrdinalAxisConfigBuilder,
-  VicYOrdinalAxisModule,
   VicYQuantitativeAxisConfig,
   VicYQuantitativeAxisConfigBuilder,
-  VicYQuantitativeAxisModule,
 } from 'libs/viz-components/src/public-api';
 import { beforeEach, cy, describe, expect, it } from 'local-cypress';
 import { VicBarsConfigBuilder } from '../bars-builder';
@@ -155,6 +152,7 @@ const barLabelColorMatchesExpectedRgb = (
     </vic-xy-chart>
   `,
   styles: [],
+  standalone: false,
 })
 class TestVerticalBarsWithLabelsComponent {
   @Input() barsConfig: BarsConfig<Datum, string>;
@@ -176,13 +174,7 @@ const mountVerticalBarsComponent = (
     .getConfig();
 
   const declarations = [TestVerticalBarsWithLabelsComponent];
-  const imports = [
-    VicChartModule,
-    VicBarsModule,
-    VicXOrdinalAxisModule,
-    VicYQuantitativeAxisModule,
-    VicXyChartModule,
-  ];
+  const imports = [VicChartModule, VicBarsModule, VicXyAxisModule];
 
   cy.mount(TestVerticalBarsWithLabelsComponent, {
     declarations,
@@ -489,6 +481,7 @@ describe('it correctly positions the vertical bar chart data labels', () => {
     </vic-xy-chart>
   `,
   styles: [],
+  standalone: false,
 })
 class TestHorizontalBarsWithLabelsComponent {
   @Input() barsConfig: BarsConfig<Datum, string>;
@@ -510,13 +503,7 @@ const mountHorizontalBarsComponent = (
   const yAxisConfig = new VicYOrdinalAxisConfigBuilder().getConfig();
 
   const declarations = [TestHorizontalBarsWithLabelsComponent];
-  const imports = [
-    VicChartModule,
-    VicBarsModule,
-    VicXQuantitativeAxisModule,
-    VicYOrdinalAxisModule,
-    VicXyChartModule,
-  ];
+  const imports = [VicChartModule, VicBarsModule, VicXyAxisModule];
 
   cy.mount(TestHorizontalBarsWithLabelsComponent, {
     declarations,
