@@ -1,10 +1,8 @@
+export * from './chart.module';
 export * from './chart/chart';
 export * from './chart/chart.component';
-export * from './chart/chart.module';
 export * from './chart/chart.token';
 export * from './chart/config/chart-builder';
 export * from './chart/config/chart-config';
 export * from './map-chart/map-chart.component';
-export * from './map-chart/map-chart.module';
 export * from './xy-chart/xy-chart.component';
-export * from './xy-chart/xy-chart.module';
