@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { DotsComponent } from './dots.component';
 import { DotsClickDirective } from './events/dots-click.directive';
@@ -7,14 +6,13 @@ import { DotsHoverDirective } from './events/dots-hover.directive';
 import { DotsInputEventDirective } from './events/dots-input.directive';
 
 @NgModule({
-  declarations: [
+  imports: [
     DotsComponent,
     DotsClickDirective,
     DotsHoverDirective,
     DotsHoverMoveDirective,
     DotsInputEventDirective,
   ],
-  imports: [CommonModule],
   exports: [
     DotsComponent,
     DotsClickDirective,
