@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @angular-eslint/prefer-standalone */
 import { Component, Input } from '@angular/core';
 import { color } from 'd3';
 import {
@@ -13,9 +14,9 @@ import { GeometryCollection, Objects, Topology } from 'topojson-specification';
 import {
   ChartConfig,
   VicChartConfigBuilder,
+  VicChartModule,
   VicGeographiesConfigBuilder,
   VicGeographiesModule,
-  VicMapChartModule,
 } from '../../../../../../../public-api';
 import {
   StateIncomePopulationYearDatum,
@@ -60,6 +61,7 @@ type TestUsMapTopology = Topology<TestMapObjects>;
     </vic-map-chart>
   `,
   styles: [],
+  standalone: false,
 })
 class TestGeographiesComponent {
   @Input() geographiesConfig: GeographiesConfig<
@@ -81,7 +83,7 @@ const mountGeographiesComponent = (
   >
 ): void => {
   const declarations = [TestGeographiesComponent];
-  const imports = [VicMapChartModule, VicGeographiesModule];
+  const imports = [VicChartModule, VicGeographiesModule];
 
   cy.mount(TestGeographiesComponent, {
     declarations,
