@@ -1,3 +1,4 @@
+/* eslint-disable @angular-eslint/prefer-standalone */
 import { Component, Input } from '@angular/core';
 import 'cypress-real-events';
 import { ascending, extent, mean, scaleLinear } from 'd3';
@@ -16,10 +17,10 @@ import {
   GeographiesHoverDirective,
   GeographiesHoverEmitTooltipData,
   VicChartConfigBuilder,
+  VicChartModule,
   VicGeographiesConfigBuilder,
   VicGeographiesModule,
   VicHtmlTooltipModule,
-  VicMapChartModule,
 } from '../../public-api';
 import { EventAction } from '../events/action';
 import {
@@ -96,6 +97,7 @@ type TestUsMapTopology = Topology<TestMapObjects>;
     </ng-template>
   `,
   styles: ['.tooltip-container { font-size: 12px; }'],
+  standalone: false,
 })
 class TestGeographiesComponent {
   @Input() geographiesConfig: GeographiesConfig<
@@ -156,11 +158,7 @@ const mountGeographiesComponent = (
   >
 ): void => {
   const declarations = [TestGeographiesComponent];
-  const imports = [
-    VicMapChartModule,
-    VicGeographiesModule,
-    VicHtmlTooltipModule,
-  ];
+  const imports = [VicChartModule, VicGeographiesModule, VicHtmlTooltipModule];
 
   cy.mount(TestGeographiesComponent, {
     declarations,
