@@ -1,10 +1,8 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { GroupedBarsComponent } from './grouped-bars.component';
 
 @NgModule({
-  declarations: [GroupedBarsComponent],
-  imports: [CommonModule],
+  imports: [GroupedBarsComponent],
   exports: [GroupedBarsComponent],
 })
 export class VicGroupedBarsModule {}
