@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { LinesClickDirective } from './events/lines-click.directive';
 import { LinesHoverMoveDirective } from './events/lines-hover-move.directive';
@@ -8,7 +7,7 @@ import { LinesMarkerClickDirective } from './events/lines-marker-click.directive
 import { LinesComponent } from './lines.component';
 
 @NgModule({
-  declarations: [
+  imports: [
     LinesComponent,
     LinesMarkerClickDirective,
     LinesClickDirective,
@@ -16,7 +15,6 @@ import { LinesComponent } from './lines.component';
     LinesInputEventDirective,
     LinesHoverDirective,
   ],
-  imports: [CommonModule],
   exports: [
     LinesComponent,
     LinesMarkerClickDirective,
