@@ -1,10 +1,8 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { QuantitativeRulesComponent } from './quantitative-rules.component';
 
 @NgModule({
-  declarations: [QuantitativeRulesComponent],
-  imports: [CommonModule],
+  imports: [QuantitativeRulesComponent],
   exports: [QuantitativeRulesComponent],
 })
 export class VicQuantitativeRulesModule {}
