@@ -1,10 +1,8 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { XyBackgroundComponent } from './xy-background.component';
 
 @NgModule({
-  declarations: [XyBackgroundComponent],
-  imports: [CommonModule],
+  imports: [XyBackgroundComponent],
   exports: [XyBackgroundComponent],
 })
 export class VicXyBackgroundModule {}
